# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SparqForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: getAssets
      tags: [assets]
      summary: Get all assets
      parameters:
        - name: brandId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create an asset record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetInput"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    put:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssetInput"
      responses:
        "200":
          description: Asset updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /creatives/{id}/generate-video:
    post:
      operationId: generateVideo
      tags: [creatives]
      summary: Generate video variants for a creative (SSE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                orientations:
                  type: array
                  items:
                    type: string
                    enum: [landscape, portrait]
      responses:
        "200":
          description: SSE stream of video generation progress
  /creatives/{id}/variants/{variantId}/audio:
    post:
      operationId: addVariantAudio
      tags: [creatives]
      summary: Generate and merge AI audio onto a video variant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: variantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudioGenerateInput"
      responses:
        "200":
          description: Variant updated with merged audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreativeVariant"
  /creatives/{id}/variants/{variantId}/audio-upload:
    post:
      operationId: uploadVariantAudio
      tags: [creatives]
      summary: Upload custom audio and merge onto a video variant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: variantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio:
                  type: string
                  format: binary
                mode:
                  type: string
                  enum: [replace, mix]
              required:
                - audio
      responses:
        "200":
          description: Variant updated with custom audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreativeVariant"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Asset:
      type: object
      properties:
        id:
          type: string
        brandId:
          type: string
        type:
          type: string
        subType:
          type: string
          nullable: true
        status:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        fileUrl:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        mimeType:
          type: string
          nullable: true
        fileSizeBytes:
          type: integer
          nullable: true
        uploadedBy:
          type: string
        approvedBy:
          type: string
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        usageCount:
          type: integer
        assetClass:
          type: string
          nullable: true
        generationRole:
          type: string
          nullable: true
        brandLayer:
          type: string
          nullable: true
        franchise:
          type: string
          nullable: true
        approvedChannels:
          type: array
          items:
            type: string
        approvedTemplates:
          type: array
          items:
            type: string
        subjectIdentityScore:
          type: number
          nullable: true
        styleStrengthScore:
          type: number
          nullable: true
        compositingOnly:
          type: boolean
        generationAllowed:
          type: boolean
        approvedForCompositing:
          type: boolean
        referencePriorityDefault:
          type: number
          nullable: true
        conflictTags:
          type: array
          items:
            type: string
        freshnessScore:
          type: number
          nullable: true
        characterIdentityNote:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - brandId
        - type
        - status
        - name
        - tags
        - uploadedBy
        - usageCount
        - approvedChannels
        - approvedTemplates
        - compositingOnly
        - generationAllowed
        - approvedForCompositing
        - conflictTags
        - createdAt
        - updatedAt
    CreateAssetInput:
      type: object
      properties:
        brandId:
          type: string
        type:
          type: string
        subType:
          type: string
          nullable: true
        name:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        fileUrl:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        mimeType:
          type: string
          nullable: true
        fileSizeBytes:
          type: integer
          nullable: true
        uploadedBy:
          type: string
        assetClass:
          type: string
          nullable: true
        generationRole:
          type: string
          nullable: true
        brandLayer:
          type: string
          nullable: true
        franchise:
          type: string
          nullable: true
        compositingOnly:
          type: boolean
        generationAllowed:
          type: boolean
        approvedForCompositing:
          type: boolean
      required:
        - brandId
        - type
        - name
        - tags
        - uploadedBy
    UpdateAssetInput:
      type: object
      properties:
        status:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        content:
          type: string
          nullable: true
        subType:
          type: string
          nullable: true
        approvedBy:
          type: string
          nullable: true
        assetClass:
          type: string
          nullable: true
        generationRole:
          type: string
          nullable: true
        brandLayer:
          type: string
          nullable: true
        franchise:
          type: string
          nullable: true
        compositingOnly:
          type: boolean
        generationAllowed:
          type: boolean
        approvedForCompositing:
          type: boolean
        subjectIdentityScore:
          type: number
          nullable: true
        styleStrengthScore:
          type: number
          nullable: true
        referencePriorityDefault:
          type: number
          nullable: true
        freshnessScore:
          type: number
          nullable: true
        characterIdentityNote:
          type: string
...
    AudioGenerateInput:
      type: object
      properties:
        type:
          type: string
          enum: [music, sfx, mute, veo_native]
        prompt:
          type: string
          nullable: true
        mode:
          type: string
          enum: [replace, mix]
        audioVolume:
          type: number
        videoVolume:
          type: number
      required:
        - type
    CreativeVariant:
      type: object
      properties:
        id:
          type: string
        creativeId:
          type: string
        platform:
          type: string
        aspectRatio:
          type: string
        rawImageUrl:
          type: string
          nullable: true
        compositedImageUrl:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        audioSource:
          type: string
          nullable: true
        audioUrl:
          type: string
          nullable: true
        mergedVideoUrl:
          type: string
          nullable: true
        caption:
          type: string
        originalCaption:
          type: string
          nullable: true
        headlineText:
          type: string
          nullable: true
        originalHeadline:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - creativeId
        - platform
        - aspectRatio
        - caption
        - status
        - createdAt
        - updatedAt
...