# lib/api-zod/src/generated/types/creativeVariant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SparqForge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreativeVariant {
  id: string;
  creativeId: string;
  platform: string;
  aspectRatio: string;
  rawImageUrl?: string | null;
  compositedImageUrl?: string | null;
  videoUrl?: string | null;
  audioSource?: string | null;
  audioUrl?: string | null;
  mergedVideoUrl?: string | null;
  caption: string;
  originalCaption?: string | null;
  headlineText?: string | null;
  originalHeadline?: string | null;
  status: string;
  createdAt: Date;
  updatedAt: Date;
}